% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_logit_funcs.R
\name{augmented_density_probit}
\alias{augmented_density_probit}
\title{augmented_density_probit}
\usage{
augmented_density_probit(Chain_count = 1, ...)
}
\value{
Computes the reweighted log-posterior distribution of the parameters of the probit model given
        the design matrix and observations. Assumes independent normal priors on the beta vectors.
        The reweighting refers to the procedure in (Neiswanger et al),
        using an unnormalized prior p(beta)^(1/Chain_count) for each subsetted data-instance.
}
\description{
augmented_density_probit
}

