% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_logit_funcs.R
\name{logit_den}
\alias{logit_den}
\title{logit_den}
\usage{
logit_den(observations, beta, design_mat, to_log = T)
}
\arguments{
\item{observations}{- A vector of length N. The observations of the logit model, i.e. usually the y_i.}

\item{beta}{- A vector of length P. The parameters of the logit model, i.e. usually the beta-vector.}

\item{to_log}{- Boolean. If True, returns the log-likelihood. If False, returns the likelihood.}

\item{design_matrix}{- A N x P matrix. The design matrix of the logit model, i.e. usually the X-matrix.}
}
\value{
(log-)The density function for the logistic regression model of the obervations given the paramaters beta.
}
\description{
logit_den
}

