% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_MCMC_chain.R
\name{MH_MCMC_chain}
\alias{MH_MCMC_chain}
\title{MH_MCMC_chain}
\usage{
MH_MCMC_chain(Iterations, target_density, proposal_sd, inital_value, ...)
}
\arguments{
\item{Iterations}{- Number of iterations for the Markov chain.}

\item{target_density}{- augmented_density function as defined.}

\item{proposal_sd}{- A number for the standard deviation used for the proposal. It is the same for all parameters.}

\item{initial_value}{- A vector of starting values for the chains, usually taken as a vector of zeroes.}
}
\value{
A matrix containing all the steps of the MH chain of dimension d x Iterations, where d is the number of parameters, i.e. length of the vector initial_value.
}
\description{
MH_MCMC_chain
}

