% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametric_Density_Product_Estimates.R
\name{parametric_implementation}
\alias{parametric_implementation}
\title{parametric_implementation}
\usage{
parametric_implementation(chain_store, burnin)
}
\arguments{
\item{chain_store}{- A list of M matrices, where M is the number of subsets.
Each matrix contains the steps of the MH chain on each subset of data.
The dimension of each matrix is Iterations x d, where d is the number of parameters.}

\item{burnin}{- A number specifying the burnin period, i.e. how many iterations will be discarded.
Usually taken as 0.2*Iterations.}
}
\value{
A matrix containing all the steps of the combined MH chain which is of dimension (Iterations - burnin)xd.
}
\description{
parametric_implementation
}

