% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_logit_funcs.R
\name{probit_den}
\alias{probit_den}
\title{probit_den}
\usage{
probit_den(observations, beta, design_mat, to_log = T)
}
\arguments{
\item{observations}{- A vector of length N. The observations of the probit model, i.e. usually the y_i.}

\item{beta}{- A vector of length P. The parameters of the probit model, i.e. usually the beta-vector.}

\item{to_log}{- Boolean. If True, returns the log-likelihood. If False, returns the likelihood.}

\item{design_matrix}{- A N x P matrix. The design matrix of the probit model, i.e. usually the X-matrix.}
}
\value{
The (log-)density function for the probit model of the obervations given the paramaters beta.
}
\description{
probit_den
}

