% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_logit_funcs.R
\name{sim_probit}
\alias{sim_probit}
\title{sim_probit}
\usage{
sim_probit(n, dimension)
}
\arguments{
\item{n}{- An integer specifying the number of observations.}

\item{dimension}{- An integer specifying the dimension of the beta vector.}
}
\value{
Generates values for the parameters, design matrix and observations of a probit model.
        Uses independent random samples for the beta parameters and the design matrix.
}
\description{
sim_probit
}

